+++ src/home/components/HowItWorks.js
@@ -8,15 +8,12 @@ const HowItWorks = () => <div className={classNames.this}>
     listItems={[
       'Farmer lists a scheduled product.',
       'Customer reserves a share of the product.',
-      'Farmer accepts payment offline and activates the share or reveals a token to the customer which is used to activate the share.',
-      'Vouchers are issued to the shareholder, who exchanges a voucher token for a product distribution during the harvest. Farmer validates the token.',
+      'Farmer accepts payment offline and activates the share or gives the customer an activation token.',
+      'Vouchers are issued to the shareholder.',
+      'Farmer validates a voucher token when product is distributed during the harvest.',
       'Vouchers can be traded and donated.',
     ]}
   />
 </div>;
 
 export default HowItWorks;
-
-// 'Farmer lists a product which is reserved by buyers as a “seasonal share”.  A share is a block of vouchers with a  schedule.',
-// 'The farmer accepts payment offline and activates the share or reveals a token to the shareholder which is used to activate the share.',
-// 'Vouchers can be traded, donated, and exchanged for product distributions during the harvest. The farmer validates a voucher with its unique token provided by the buyer.',